+++ core/src/main/java/org/apache/carbondata/core/mutate/data/BlockMappingVO.java
@@ -30,6 +30,8 @@ public class BlockMappingVO {
 
   private Map<String, RowCountDetailsVO> completeBlockRowDetailVO;
 
+  // This map will help us to finding the segment id from the block path.
+  // key is 'blockpath' and value is 'segmentId'
   private Map<String, String> blockToSegmentMapping;
 
   public void setCompleteBlockRowDetailVO(Map<String, RowCountDetailsVO> completeBlockRowDetailVO) {
