+++ core/src/main/java/org/apache/carbondata/core/statusmanager/SegmentUpdateStatusManager.java
@@ -176,10 +176,6 @@ public class SegmentUpdateStatusManager {
 
   }
 
-  public Map<String, SegmentUpdateDetails> getBlockAndDetailsMap() {
-    return blockAndDetailsMap;
-  }
-
   /**
    * Returns the LoadMetadata Details
    * @return
