+++ index/secondary-index/src/test/scala/org/apache/carbondata/spark/testsuite/secondaryindex/TestIndexModelForORFilterPushDown.scala
@@ -17,9 +17,9 @@
 package org.apache.carbondata.spark.testsuite.secondaryindex
 
 import org.apache.carbondata.core.util.CarbonProperties
+import org.apache.carbondata.spark.testsuite.secondaryindex.TestSecondaryIndexUtils
+.isFilterPushedDownToSI;
 import org.apache.spark.sql.Row
-import org.apache.spark.sql.execution.SparkPlan
-import org.apache.spark.sql.secondaryindex.joins.BroadCastSIFilterPushJoin
 import org.apache.spark.sql.test.util.QueryTest
 import org.scalatest.BeforeAndAfterAll
 
@@ -234,21 +234,4 @@ class TestIndexModelForORFilterPushDown extends QueryTest with BeforeAndAfterAll
   override def afterAll: Unit = {
     dropTables
   }
-
-  /**
-   * Method to check whether the filter is push down to SI table or not
-   *
-   * @param sparkPlan
-   * @return
-   */
-  def isFilterPushedDownToSI(sparkPlan: SparkPlan): Boolean = {
-    var isValidPlan = false
-    sparkPlan.transform {
-      case broadCastSIFilterPushDown: BroadCastSIFilterPushJoin =>
-        isValidPlan = true
-        broadCastSIFilterPushDown
-    }
-    isValidPlan
-  }
-
 }
